FEATURE: Done import excel file for products list

+++ src/main/java/com/example/swp/responses/ProductResponse.java
@@ -18,7 +18,7 @@ public class ProductResponse {
     @JsonProperty(value = "barcode")
     private String barcode;
     @JsonProperty(value = "quantity")
-    private Long quantity;
+    private int quantity;
     @JsonProperty(value = "price_processing")
     private double priceProcessing;
     @JsonProperty(value = "price_stone")
